compile/requirements.txt
yt-dlp
opencv-python
tqdm
numba
numpy
cupy-cuda12x
